Refactor middleware and rule engine
